requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --override overrides.txt -o requirements.txt pyproject.toml
aiohttp==3.9.5
    # via web3
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
ape-alchemy==0.7.2
ape-foundry==0.7.4
ape-vyper==0.7.0
asttokens==2.4.1
    # via
    #   stack-data
    #   vyper
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
base58==1.0.3
    # via
    #   py-cid
    #   py-multihash
bitarray==2.9.2
    # via eth-account
cached-property==1.5.2
    # via
    #   py-ecc
    #   py-evm
cbor2==5.6.3
    # via vyper
certifi==2024.2.2
    # via requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via eth-ape
commonmark==0.9.1
    # via rich
cryptography==42.0.5
    # via pyjwt
cytoolz==0.12.3
    # via eth-utils
dataclassy==0.11.1
    # via eip712
decorator==5.1.1
    # via ipython
deprecated==1.2.14
    # via pygithub
eip712==0.2.5
    # via eth-ape
eth-abi==4.2.1
    # via
    #   eip712
    #   eth-account
    #   eth-ape
    #   eth-tester
    #   web3
eth-account==0.10.0
    # via
    #   eip712
    #   eth-ape
    #   eth-tester
    #   web3
eth-ape==0.7.11
    # via
    #   ape-alchemy
    #   ape-foundry
    #   ape-vyper
eth-bloom==3.0.1
    # via py-evm
eth-hash==0.7.0
    # via
    #   eip712
    #   eth-bloom
    #   eth-pydantic-types
    #   eth-tester
    #   eth-utils
    #   trie
    #   web3
eth-keyfile==0.6.1
    # via eth-account
eth-keys==0.4.0
    # via
    #   eth-account
    #   eth-keyfile
    #   eth-tester
    #   py-evm
eth-pydantic-types==0.1.0
    # via
    #   ape-alchemy
    #   ape-foundry
    #   eth-ape
    #   ethpm-types
    #   evm-trace
eth-rlp==0.3.0
    # via eth-account
eth-tester==0.9.1b2
    # via web3
eth-typing==3.5.2
    # via
    #   eip712
    #   eth-abi
    #   eth-ape
    #   eth-keys
    #   eth-pydantic-types
    #   eth-utils
    #   py-ecc
    #   py-evm
    #   web3
eth-utils==2.3.1
    # via
    #   eip712
    #   eth-abi
    #   eth-account
    #   eth-ape
    #   eth-keyfile
    #   eth-keys
    #   eth-pydantic-types
    #   eth-rlp
    #   eth-tester
    #   ethpm-types
    #   evm-trace
    #   py-ecc
    #   py-evm
    #   rlp
    #   trie
    #   web3
ethpm-types==0.6.11
    # via
    #   ape-alchemy
    #   ape-foundry
    #   ape-vyper
    #   eth-ape
evm-trace==0.1.3
    # via
    #   ape-alchemy
    #   ape-foundry
    #   eth-ape
evmchains==0.0.7
    # via eth-ape
executing==2.0.1
    # via stack-data
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
hexbytes==0.3.1
    # via
    #   ape-foundry
    #   eip712
    #   eth-account
    #   eth-pydantic-types
    #   eth-rlp
    #   trie
    #   web3
idna==3.7
    # via
    #   requests
    #   yarl
ijson==3.2.3
    # via eth-ape
importlib-metadata==7.1.0
    # via
    #   eth-ape
    #   vyper
iniconfig==2.0.0
    # via pytest
ipython==8.24.0
    # via eth-ape
jedi==0.19.1
    # via ipython
jsonschema==4.21.1
    # via web3
jsonschema-specifications==2023.12.1
    # via jsonschema
lazyasd==0.1.4
    # via eth-ape
lru-dict==1.2.0
    # via
    #   py-evm
    #   web3
matplotlib-inline==0.1.7
    # via ipython
morphys==1.0
    # via
    #   py-cid
    #   py-multibase
    #   py-multicodec
    #   py-multihash
msgspec==0.18.6
    # via evm-trace
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via
    #   py-ecc
    #   py-evm
numpy==1.26.4
    # via pandas
packaging==23.2
    # via
    #   eth-ape
    #   pytest
    #   vvm
    #   vyper
pandas==1.5.3
    # via eth-ape
parsimonious==0.9.0
    # via eth-abi
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pluggy==1.5.0
    # via
    #   eth-ape
    #   pytest
prompt-toolkit==3.0.43
    # via ipython
protobuf==5.26.1
    # via web3
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
py-cid==0.3.0
    # via ethpm-types
py-ecc==6.0.0
    # via py-evm
py-evm==0.7.0a4
    # via
    #   eth-tester
    #   evm-trace
py-geth==4.4.0
    # via
    #   eth-ape
    #   web3
py-multibase==1.0.3
    # via py-cid
py-multicodec==0.2.1
    # via py-cid
py-multihash==0.2.3
    # via py-cid
pycparser==2.22
    # via cffi
pycryptodome==3.20.0
    # via
    #   eth-hash
    #   eth-keyfile
    #   vyper
pydantic==2.7.1
    # via
    #   eth-ape
    #   eth-pydantic-types
    #   ethpm-types
    #   evm-trace
    #   evmchains
    #   pydantic-settings
pydantic-core==2.18.2
    # via pydantic
pydantic-settings==2.2.1
    # via eth-ape
pyethash==0.1.27
    # via py-evm
pygithub==1.59.1
    # via eth-ape
pygments==2.17.2
    # via
    #   ipython
    #   rich
pyjwt==2.8.0
    # via pygithub
pynacl==1.5.0
    # via pygithub
pytest==7.4.4
    # via eth-ape
python-baseconv==1.2.2
    # via py-multibase
python-dateutil==2.9.0.post0
    # via
    #   eth-ape
    #   pandas
python-dotenv==1.0.1
    # via pydantic-settings
pytz==2024.1
    # via pandas
pyunormalize==15.1.0
    # via web3
pyyaml==6.0.1
    # via eth-ape
referencing==0.35.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.4.28
    # via parsimonious
requests==2.31.0
    # via
    #   ape-alchemy
    #   eth-ape
    #   ethpm-types
    #   pygithub
    #   vvm
    #   web3
rich==12.6.0
    # via eth-ape
rlp==3.0.0
    # via
    #   eth-account
    #   eth-rlp
    #   eth-tester
    #   py-evm
    #   trie
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
safe-pysha3==1.0.4
    # via eth-hash
semantic-version==2.10.0
    # via
    #   eth-tester
    #   py-geth
six==1.16.0
    # via
    #   asttokens
    #   py-multibase
    #   py-multicodec
    #   py-multihash
    #   python-dateutil
sortedcontainers==2.4.0
    # via trie
sqlalchemy==2.0.29
    # via eth-ape
stack-data==0.6.3
    # via ipython
toolz==0.12.1
    # via cytoolz
tqdm==4.66.2
    # via
    #   ape-vyper
    #   eth-ape
traitlets==5.14.3
    # via
    #   eth-ape
    #   ipython
    #   matplotlib-inline
trie==2.2.0
    # via py-evm
typing-extensions==4.11.0
    # via
    #   eth-pydantic-types
    #   eth-typing
    #   ethpm-types
    #   ipython
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   web3
urllib3==2.2.1
    # via
    #   eth-ape
    #   requests
varint==1.0.2
    # via
    #   py-multicodec
    #   py-multihash
vvm==0.2.0
    # via ape-vyper
vyper==0.3.10
watchdog==3.0.0
    # via eth-ape
wcwidth==0.2.13
    # via prompt-toolkit
web3==6.16.0
    # via
    #   ape-alchemy
    #   ape-foundry
    #   eth-ape
websockets==12.0
    # via web3
wheel==0.43.0
    # via vyper
wrapt==1.16.0
    # via deprecated
yarl==1.9.4
    # via
    #   aiohttp
    #   ape-foundry
zipp==3.18.1
    # via importlib-metadata
